Fix curve fuzzers
